FOA-Admin-Module1/requirements.txt
flask_sqlalchemy
flask_migrate
flask_restful
psycopg2-binary 
PyJWT
flask_cors
requests
sqlalchemy-utils
cryptography
pycryptodome